Fix issue where scrolling was not disabled at page load.
After a recent update to Google Calendar, the scrolling was still working right when the page loaded.
However, the scrolling stopped working after the user changed to a different view.

The issue was that the data-viewfamily mutation no longer occurs when the page is loaded.
Now, the data-viewfamily mutation only occurs when the user changes views.
Thus, we need to disable scrolling once when the page is initially loaded.