Fix issue where changing months re-enabled scroll wheel.
To do this, we block scroll events on the "main" div instead of the "grid" div.

With the latest version of Google Calendar, the "grid" div is replaced whenever the month is changed.
Specifically, the grid is replaced once the new events have loaded.
Unfortunately, the mutation event occurs before the events have loaded (and before the grid has been replaced).